[ERM-5124] тип свойств, прописанных в DoubleGis.Erm.Platform.Model.Entities.Enums.Infrastructure.PropertiesToEnumMapper, изменен на соответствующие enum'ы
git-tfs-id: [http://uk-tfs02:8080/tfs/DefaultCollection]$/Erm.Core/Dev/Features/EFCF/CompositionRoots;C25688

+++ Source/Tests/2Gis.Erm.Tests.Integration.InProc/Suite/Concrete/BLFlex/Operations/Concrete/OrderProcessingRequests/RequestStateApplicationServiceTest.cs
@@ -24,7 +24,7 @@ namespace DoubleGis.Erm.Tests.Integration.InProc.Suite.Concrete.BLFlex.Operation
 
         protected override FindSpecification<OrderProcessingRequest> ModelEntitySpec
         {
-            get { return base.ModelEntitySpec && new FindSpecification<OrderProcessingRequest>(x => x.State != (int)OrderProcessingRequestState.Undefined); }
+            get { return base.ModelEntitySpec && new FindSpecification<OrderProcessingRequest>(x => x.State != OrderProcessingRequestState.Undefined); }
         }
 
         protected override OrdinaryTestResult ExecuteWithModel(OrderProcessingRequest modelEntity)
